sites_task_test: Catch initial periodic sync
This fixes the occasional failure where the SyncAll could come before the
SyncSite command triggered by mntchng. This does not technically remove the race
condition, because if the periodic sync interval is low or the PC is very slow,
we could get another SyncAll command, before catching the SyncSite command.